+++ tests/dem/distribution_normal.cc
@@ -52,17 +52,11 @@ test()
   DEMSolverParameters<dim> dem_parameters;
 
   // Defining simulation general parameters
-  dem_parameters.insertion_info.x_min                                = -0.5;
-  dem_parameters.insertion_info.y_min                                = -0.5;
-  dem_parameters.insertion_info.z_min                                = -0.05;
-  dem_parameters.insertion_info.x_max                                = 0.5;
-  dem_parameters.insertion_info.y_max                                = 0.5;
-  dem_parameters.insertion_info.z_max                                = 0.05;
-  dem_parameters.insertion_info.axis_0                               = 0;
-  dem_parameters.insertion_info.axis_1                               = 1;
-  dem_parameters.insertion_info.axis_2                               = 2;
-  dem_parameters.insertion_info.inserted_this_step                   = 100;
-  dem_parameters.insertion_info.distance_threshold                   = 2;
+  dem_parameters.insertion_info.insertion_box_point_1 = {-0.5, -0.5, -0.05};
+  dem_parameters.insertion_info.insertion_box_point_2 = {0.5, 0.5, 0.05};
+  dem_parameters.insertion_info.direction_sequence    = {0, 1, 2};
+  dem_parameters.insertion_info.inserted_this_step    = 100;
+  dem_parameters.insertion_info.distance_threshold    = 2;
   dem_parameters.lagrangian_physical_properties.particle_type_number = 1;
   dem_parameters.lagrangian_physical_properties.distribution_type.push_back(
     Parameters::Lagrangian::SizeDistributionType::normal);